Extract responsibilities from router into discoveryResultsFetcher, messageSerializer, responseGenerator.
